Debugging stuff

+++ tests/forward_pass_logit_checker.py
@@ -378,7 +378,7 @@ def main(config, test_args):  # pylint: disable=W0621
     """Comparing maxtext model with HF model on-the-fly"""
     if test_args.hf_model_path == "":
       raise ValueError("run_hf_model requires hf_model_path")
-    hf_model = AutoModelForCausalLM.from_pretrained(test_args.hf_model_path, dtype=torch.bfloat16)
+    hf_model = AutoModelForCausalLM.from_pretrained(test_args.hf_model_path, dtype=torch.float32)
     tokenizer = AutoTokenizer.from_pretrained(test_args.hf_model_path)
     if "Llama-3.1" in test_args.hf_model_path:
       tokenizer.pad_token = tokenizer.eos_token
